generalize _seed_id_map and move to obspy.core.util.misc
